+++ lib/url.c
@@ -411,7 +411,7 @@ CURLcode Curl_close(struct SessionHandle *data)
           if(data == (struct SessionHandle *) curr->ptr) {
             fprintf(stderr,
                     "MAJOR problem we %p are still in send pipe for %p done %d\n",
-                    data, connptr, connptr->bits.done);
+                    data, connptr, (int)connptr->bits.done);
           }
         }
       }
@@ -421,7 +421,7 @@ CURLcode Curl_close(struct SessionHandle *data)
           if(data == (struct SessionHandle *) curr->ptr) {
             fprintf(stderr,
                     "MAJOR problem we %p are still in recv pipe for %p done %d\n",
-                    data, connptr, connptr->bits.done);
+                    data, connptr, (int)connptr->bits.done);
           }
         }
       }
@@ -431,7 +431,7 @@ CURLcode Curl_close(struct SessionHandle *data)
           if(data == (struct SessionHandle *) curr->ptr) {
             fprintf(stderr,
                     "MAJOR problem we %p are still in pend pipe for %p done %d\n",
-                    data, connptr, connptr->bits.done);
+                    data, connptr, (int)connptr->bits.done);
           }
         }
       }
